requirements.txt
ollama>=0.4.0
edge-tts>=6.1.9
pygame>=2.5.0
python-dotenv>=1.0.0
rich>=13.7.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.14
